service's setter is added into  command

+++ src/main/java/ru/multa/entia/credential/impl/data/manager/command/DeleteUsrByIdManagerCommand.java
@@ -1,30 +1,45 @@
 package ru.multa.entia.credential.impl.data.manager.command;
 
+import lombok.Setter;
 import org.bson.types.ObjectId;
 import ru.multa.entia.credential.api.data.manager.item.ManagerDatum;
 import ru.multa.entia.credential.api.data.usr.Usr;
 import ru.multa.entia.credential.api.data.usr.UsrService;
 import ru.multa.entia.credential.impl.data.manager.item.DefaultManagerDatum;
+import ru.multa.entia.results.api.repository.CodeRepository;
 import ru.multa.entia.results.api.result.Result;
+import ru.multa.entia.results.impl.repository.DefaultCodeRepository;
 import ru.multa.entia.results.impl.result.DefaultResultBuilder;
 
 import java.util.function.Consumer;
 
-// TODO: impl service setter
 public class DeleteUsrByIdManagerCommand extends AbstractManagerCommand{
-    private final UsrService usrService;
+    public enum Code {
+        SERVICE_IS_ABSENCE
+    }
+
+    private static final CodeRepository CR = DefaultCodeRepository.getDefaultInstance();
+    static {
+        CR.update(Code.SERVICE_IS_ABSENCE, "credential:command.delete-usr-by-id:service-is-absence");
+    }
+
     private final ObjectId id;
 
+    @Setter
+    private UsrService usrService;
+
     public DeleteUsrByIdManagerCommand(final Consumer<Result<ManagerDatum>> consumer,
-                                       final UsrService usrService,
                                        final ObjectId id) {
         super(consumer);
-        this.usrService = usrService;
         this.id = id;
     }
 
     @Override
     protected Result<ManagerDatum> executeService() {
+        if (usrService == null) {
+            return DefaultResultBuilder.<ManagerDatum>fail(CR.get(Code.SERVICE_IS_ABSENCE));
+        }
+
         Result<Usr> result = usrService.deleteById(id);
         return result.ok()
                 ? DefaultResultBuilder.<ManagerDatum>ok(DefaultManagerDatum.builder().build())
